requirements.txt
fastapi==0.121.1
PyYaml==6.0.3
pydantic_yaml==1.4.0
email-validator==2.3.0
dotenv==0.9.9
redis==7.0.1
jinja2==3.1.6
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
docxtpl==0.20.1
python-docx==1.2.0
weasyprint==66.0
docx2pdf==0.1.8
rocketmq-client-python==2.0.0